+++ .gitattributes
@@ -47,7 +47,7 @@ PC/clinic/*.h               linguist-generated=true
 Python/clinic/*.h           linguist-generated=true
 Python/importlib.h          linguist-generated=true
 Python/importlib_external.h linguist-generated=true
-Include/Python-ast.h        linguist-generated=true
+Include/internal/pycore_ast.h   linguist-generated=true
 Python/Python-ast.c         linguist-generated=true
 Include/opcode.h            linguist-generated=true
 Python/opcode_targets.h     linguist-generated=true
