+++ Doc/whatsnew/3.10.rst
@@ -1383,5 +1383,13 @@ Removed
   it could not be used, because the ``symtable.h`` header file was excluded
   from the limited C API.
 
-  The Python :mod:`symtable` module remains available and is unchanged.
+  Use Python :mod:`symtable` module instead.
+  (Contributed by Victor Stinner in :issue:`43244`.)
+
+* Remove ``ast.h``, ``asdl.h``, and ``Python-ast.h`` header files.
+  These functions were undocumented and excluded from the limited C API.
+  Most names defined by these header files were not prefixed by ``Py`` and so
+  could create names conflicts. For example, ``Python-ast.h`` defined a
+  ``Yield`` macro which was conflict with the ``Yield`` name used by the
+  Windows ``<winbase.h>`` header. Use the Python :mod:`ast` module instead.
   (Contributed by Victor Stinner in :issue:`43244`.)
