+++ Misc/NEWS.d/next/C API/2021-03-19-12-56-11.bpo-43244.VuIyOD.rst
@@ -1,0 +1,8 @@
+Remove ``ast.h``, ``asdl.h``, and ``Python-ast.h`` header files.
+These functions were undocumented and excluded from the limited C API.
+Most names defined by these header files were not prefixed by ``Py`` and so
+could create names conflicts. For example, ``Python-ast.h`` defined a ``Yield``
+macro which was conflict with the ``Yield`` name used by the Windows
+``<winbase.h>`` header.
+Use the Python :mod:`ast` module instead.
+Patch by Victor Stinner.
