+++ Python/asdl.c
@@ -1,5 +1,5 @@
 #include "Python.h"
-#include "asdl.h"
+#include "pycore_asdl.h"
 
 GENERATE_ASDL_SEQ_CONSTRUCTOR(generic, void*);
 GENERATE_ASDL_SEQ_CONSTRUCTOR(identifier, PyObject*);
