+++ Python/bltinmodule.c
@@ -2,9 +2,8 @@
 
 #include "Python.h"
 #include <ctype.h>
-#include "ast.h"
-#undef Yield   /* undefine macro conflicting with <winbase.h> */
 #include "pycore_ast.h"           // _PyAST_Validate()
+#undef Yield   /* undefine macro conflicting with <winbase.h> */
 #include "pycore_object.h"        // _Py_AddToAllObjects()
 #include "pycore_pyerrors.h"      // _PyErr_NoMemory()
 #include "pycore_pystate.h"       // _PyThreadState_GET()
